% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cgaim.R
\name{print.cgaim}
\alias{print.cgaim}
\title{Print a cgaim object}
\usage{
\method{print}{cgaim}(x, ...)
}
\arguments{
\item{x}{A \code{cgaim} object.}

\item{...}{For compatibility with the default \code{print} method. Unused
at the moment.}
}
\value{
Function called for its side effect of printing a \code{cgaim} object. Returns no value.
}
\description{
Default method to print the results from a call to \code{cgaim}. 
Conveniently prints the formula, \code{beta} and \code{alpha} coefficients
   as well as the residual sum of squares.
}
\seealso{
The main fitting function \code{\link{cgaim}}.
}
