% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_setup.R
\name{alpha.control}
\alias{alpha.control}
\title{Alpha control}
\usage{
alpha.control(Cmat = NULL, bvec = NULL, alpha.start = NULL,
  init.type = "regression", norm.type = "1", solver = "quadprog",
  ctol = 0.001, qp_pars = list())
}
\arguments{
\item{Cmat}{Overall constraint matrix for alpha coefficients.
Number of columns must match the total number of alpha coefficients. 
Constraints on individual indices can also be conveniently specified
through the \code{\link{g}} function in the formula.}

\item{bvec}{Vector containing boundary values for the constraints. If only
Cmat is provided, the default is a vector of zeros.}

\item{alpha.start}{A vector or list of starting alpha values. If \code{NULL},
starting values are generated internally.}

\item{init.type}{If \code{alpha.start = NULL}, the type of starting values
to generate. If \code{init.type = "regression"} (the default), the
starting values are generated by regressing the index design matrix on
the response. If \code{init.type = "random"}, the starting values are
generated randomly, ensuring they fall in the feasible region. In the
latter case, it might be wise to estimate the model several times with
different starting values to keep the best fit.}

\item{norm.type}{Character giving the type of normalization applied on
each alpha_j vector for it to have norm 1. 
See \code{\link[base:norm]{norm}}.}

\item{solver}{The quadratic programming solver to use. Implemented
solvers are "quadprog" and "osqp" (see corresponding packages).
The former is a bit slower but gives more accurate solutions.}

\item{ctol}{Tolerance value to be added to \code{bvec} giving some margin
on the constraints. Useful for cases in which rounding errors result
in alphas slightly violating the constraints.}

\item{qp_pars}{A named list of parameters to be passed to the QP function.
See help pages for the corresponding functions.}
}
\description{
Parameters for fine tuning of alpha coefficients estimation.
}
