% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.cgaim.R
\name{boot.cgaim}
\alias{boot.cgaim}
\title{Boostrap CGAIM}
\usage{
boot.cgaim(object, boot.type = c("residuals", "wild", "pairs"),
  bsamples = NULL, B = 100, l = 1, nc = 1)
}
\arguments{
\item{object}{A \code{cgaim} object.}

\item{boot.type}{The type of bootstrap to perform. Currently
available type are \code{"residuals"}, \code{"wild"} and 
\code{"pairs"}. See details}

\item{bsamples}{A numerical matrix of observation indices specifying 
bootstrap samples.
Rows indicate observations and columns bootstrap samples.
If \code{NULL} (the default), samples are generated internally.}

\item{B}{Number of bootstrap samples to generate when \code{bsamples = NULL}.}

\item{l}{Block length for block-bootstrap. Samples are generated by
resampling block of observation of length \code{l}. The classical
bootstrap corresponds to \code{l = 1} (the default).}

\item{nc}{Number of cores to be used for parallel computing.

This function fits \code{cgaim} on bootstrap samples of the original dataset.
It is used by the \code{\link{confint.cgaim}} function, but can also be
called directly to generate various statistics.

Two types of bootstrap are currently implemented. \code{"residuals"} 
(the default) resamples residuals that are added to fitted values by the
model. Boostrap replications are then computed by fitting \code{cgaim} 
using these newly generated y values with the original x. \code{"wild"} is
similar except that residuals are multiplied by a random draws from a
standard normal distribution before being added to fitted values. In
\code{boot.type = "pairs"} pairs of y and x are resampled to create 
bootstrap samples.

Indices for bootstrap samples can either be given by the user through
\code{bsamples} or generated internally. In the former case, 
the columns of \code{bsamples} indicate the number of replications and
the rows should match the original number observations. Internally
generated bootstrap samples are controlled by the number of replications
\code{B} and block length \code{l} if one wants to use block bootstrap.
The latter is particularly recommended for time series data.

As fitting a large number of \code{cgaim} models can be computationally
intense, the function can be run in parallel, using the 
\code{\link{doParallel}} package. This can be done by setting the
argument \code{nc} to a value greater than 1, controlling the number
of cores used in parallelization.}
}
\description{
Generates bootstrap replicates of a \code{cgaim} object.
}
