% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_constraints.R
\name{build_constraints}
\alias{build_constraints}
\title{Common constraints}
\usage{
build_constraints(p, first = 1, sign = 0, monotone = 0, convex = 0)
}
\arguments{
\item{p}{The number of variables.}

\item{first}{Indicates sign constraint for first coefficient.
Recommended for identifiability purposes but overriden by any 
other constraint.}

\item{sign}{Constrained sign for each coefficient. \code{0}: no constraint,}

\item{monotone}{Monotonicity constraint. \code{0}: no constraint, \code{-1}:
decreasing coefficients and \code{1}: increasing coefficients.}

\item{convex}{Convexity constraint. \code{0}: no constraint, \code{-1}:
convex coefficients and \code{1}: concave coefficients.}
}
\description{
Build a constraint matrix from common simple constraints.
}
\details{
For monotonicity and convexity / concavity, the function assumes the 
coefficients are ordered. For instance, for increasing monotone coefficients,
the the first one will be lower than the second, which be lower than the
third and so on.

If a different value than -1, 0 or 1 is passed to an argument, 
the function uses its sign.
}
