% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_functions.R
\name{build_constraints}
\alias{build_constraints}
\title{Common constraints}
\usage{
build_constraints(nvars, monotone = 0, sign.const = 0, first.const = 0)
}
\arguments{
\item{nvars}{A vector giving, for each index, its number of variables.}

\item{monotone}{A vector the same length as \code{nvars} giving
monotonicity constraints. \code{0} means no constraint, \code{-1}
decreasing coefficients and \code{1} increasing coefficients.}

\item{sign.const}{A vector the same length as \code{nvars} giving
signs constraints. \code{0} means no constraint, \code{-1}
means that, for the corresponding index, all \eqn{\alpha <= 0} and
\code{-1} all \eqn{\alpha >= 0}.}

\item{first.const}{A vector indicating a sign constraint for first
coefficient of each index. Recommended for indentifiability purposes but
overriden by \code{sign.const}.}
}
\description{
Build a constraint matrix from common simple constraints.
}
