% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.cgaim.R
\name{confint.cgaim}
\alias{confint.cgaim}
\title{Confidence intervals}
\usage{
\method{confint}{cgaim}(object, parm, level = 0.95, type = c("normal",
  "bootstrap"), B = 100, ...)
}
\arguments{
\item{object}{A \code{cgaim} object.}

\item{parm}{The type of parameters for which to get confidence intervals.}

\item{level}{The level of confidence intervals. Default to 95\%.}

\item{type}{A character vector giving the type(s) of confidence intervals
to compute. \code{"normal"} compute classical confidence intervals
based on normal approximation. The others are all based on bootstrap.
\code{"boot.pct"} computes intervals as percentiles of the bootstrap
distribution. \code{"boot.t"} computes studentized bootstrap intervals.
\code{"boot.bca"} (still experimental, use at your own risk) 
computes bias-corrected and accelerated bootstrap
intervals. Can include several types.
CURRENTLY, ONLY "boot.pct" WORKS}

\item{...}{Additional parameters to be passed to the parallel function. See
\code{\link[parallel]{mclapply}} or \code{\link[parallel]{parLapply}}.}
}
\value{
A named list with elements \code{alpha}, \code{beta} and \code{g}.
   Each is a list containing as elements the different types of 
   intervals given in \code{type}. In addition, contains the generated
   bootstrap values as well as the bias and acceleration computed if
   \code{"boot.bca"} is in \code{type}.
}
\description{
Computes confidence intervals for the CGAIM parameters.
}
\references{
Pya, N., Wood, S.N., 2015. Shape constrained additive models. 
   Stat. Comput. 25, 543–559. 
  Wood, S.N., 2017. Generalized Additive Models: An Introduction with R, 
    2nd ed, Texts in Statistical Science. Chapman and Hall/CRC.
  DiCiccio, T.J., Efron, B., 1996. Bootstrap Confidence Intervals. 
    Statistical Science 11, 189–212.
  Carpenter, J., Bithell, J., 2000. Bootstrap confidence intervals: 
    when, which, what? A practical guide for medical statisticians. 
    Statistics in Medicine 19, 1141–1164.
}
