% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{confint.cgaim}
\alias{confint.cgaim}
\title{Confidence intervals}
\usage{
\method{confint}{cgaim}(
  object,
  parm,
  level = 0.95,
  type = c("boot.pct", "boot.t", "boot.bca", "normal"),
  boot.type = c("residuals", "pairs"),
  bsamples = NULL,
  B = 100,
  l = 1,
  m = 20,
  applyFun = c("lapply", "mclapply", "parLapply"),
  cl = NULL,
  ...
)
}
\arguments{
\item{object}{A \code{cgaim} object.}

\item{parm}{A numeric or character vector indicating the terms for which
to compute the confidence intervals. Indicates only indices and 
covariates, not individual alpha coefficients. If missing, includes
all terms.}

\item{level}{The level of confidence intervals. Default to 95\%.}

\item{type}{A character vector giving the type(s) of confidence intervals
to compute. \code{"normal"} compute classical confidence intervals
based on normal approximation. The others are all based on bootstrap.
\code{"boot.pct"} computes intervals as percentiles of the bootstrap
distribution. \code{"boot.t"} computes studentized bootstrap intervals.
\code{"boot.bca"} (still experimental, use at your own risk) 
computes bias-corrected and accelerated bootstrap
intervals. Can include several types.
CURRENTLY, ONLY "boot.pct" WORKS}

\item{boot.type}{Character indicating the type of resampling for bootstrap.
If \code{boot.type = "residuals"} (the default), residuals are
resampled and added to fitted valued to create a new bootstrapped 
response sample. If \code{boot.type = "pairs"}, new samples are
created by resampling pairs, i.e. both response and predictors.}

\item{bsamples}{A n x B matrix of observation indices used for user-specified 
bootstrap samples. If \code{NULL}, samples are generated internally.}

\item{B}{Number of bootstrap samples to generate if \code{bsamples = NULL}.}

\item{l}{Block length for block-bootstrap. Samples are generated by
resampling block of observation of length \code{l}. The classical
bootstrap corresponds to \code{l = 1} (the default).}

\item{m}{If \code{type} includes \code{"boot.bca"}, the number of jackknife 
blocks used to compute the acceleration parameter.}

\item{applyFun}{The 'apply' function used to estimate the CGAIM on each
bootstrap sample. Used to parallelize computation. Note that 
\code{applyFun = "mclapply"} does not work on windows. Default to
\code{"lapply"}, i.e. without parallelization.}

\item{cl}{A cluster object to be used when 
\code{applyFun = "parLapply"}. See \code{\link[parallel]{parLapply}}.}

\item{...}{Additional parameters to be passed to the parallel function. See
\code{\link[parallel]{mclapply}} or \code{\link[parallel]{parLapply}}.}
}
\value{
A named list with elements \code{alpha}, \code{beta} and \code{g}.
   Each is a list containing as elements the different types of 
   intervals given in \code{type}. In addition, contains the generated
   bootstrap values as well as the bias and acceleration computed if
   \code{"boot.bca"} is in \code{type}.
}
\description{
Computes confidence intervals for the CGAIM parameters.
}
\references{
Pya, N., Wood, S.N., 2015. Shape constrained additive models. 
   Stat. Comput. 25, 543–559. 
  Wood, S.N., 2017. Generalized Additive Models: An Introduction with R, 
    2nd ed, Texts in Statistical Science. Chapman and Hall/CRC.
  DiCiccio, T.J., Efron, B., 1996. Bootstrap Confidence Intervals. 
    Statistical Science 11, 189–212.
  Carpenter, J., Bithell, J., 2000. Bootstrap confidence intervals: 
    when, which, what? A practical guide for medical statisticians. 
    Statistics in Medicine 19, 1141–1164.
}
