% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_functions.R
\name{g}
\alias{g}
\title{Defining indices in CGAIM}
\usage{
g(..., label = term[1], constraints = list(), bs = "tp", sOptions = list())
}
\arguments{
\item{...}{A list of variables on which the index is based. May contain vectors and matrices.}

\item{label}{A character used to give a name to the index. Useful for
displaying the results.}

\item{constraints}{A list of common constraints for the index. See 
\code{\link{build_constraints}} for built-in constraints. For
other constraints, see \code{\link{alpha.setup}}.}

\item{bs}{The type of basis functions for the ridge function. Used to
give shape constraints. The bases allowed are expressed as in the
\code{scam} (\code{\link[scam]{shape.constrained.smooth.terms}}) and
\code{mgcv} (\code{\link[mgcv]{smooth.terms}}) packages.}

\item{sOptions}{A named list of options to be passed to \code{\link[mgcv]{s}}
for the ridge function smoothing.}
}
\description{
Function used to define indices whithin a \code{cgaim} formula.
}
