% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cgaim.R
\name{plot.cgaim}
\alias{plot.cgaim}
\title{Plot CGAIM terms}
\usage{
\method{plot}{cgaim}(x, select = NULL, ci = NULL, ci.plot = c("polygon",
  "lines"), ci.args = list(), add = FALSE, xcenter = FALSE,
  xscale = FALSE, yshift = FALSE, yscale = FALSE, ...)
}
\arguments{
\item{x}{A \code{cgaim} object.}

\item{select}{A numeric or character vector indicating which terms
to plot.}

\item{ci}{An object returned by a call to \code{\link{confint.cgaim}}. If
\code{NULL}, no confidence interval is drawn.}

\item{ci.plot}{Whether to plot the confidence intervals as shaded areas
\code{ci.plot = "polygon"} or as lines \code{ci.plot = "lines"}.}

\item{ci.args}{Additional arguments to be passed to the function used
to draw confidence interval. Either \code{\link[graphics]{polygon}} or
\code{\link[graphics]{lines}}.}

\item{...}{Additional graphical parameters for the drawn function. See
\code{\link[graphics]{par}}.
yshift and yscale can take intercept and beta if set to NULL}
}
\description{
Plot method for the ridge and smooth terms of a \code{cgaim} object. If
   provided, may also plot confidence intervals.
}
\details{
At the moment, only plots ridge and smooth functions. When several terms
   are plotted, waits for user response before displaying the next one.
}
