% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.cgaim}
\alias{predict.cgaim}
\title{Predictions from a fitted CGAIM object}
\usage{
\method{predict}{cgaim}(object, newdata, type = c("response", "terms",
  "indices"), select = NULL, na.action = "na.pass", ...)
}
\arguments{
\item{object}{A \code{gaim} object.}

\item{newdata}{A list or data.frame containing the new data to predict.
If missing, fitted values from the model are returned.}

\item{type}{A character indicating the type of prediction to return.
When \code{type = "response"}, the predicted response is returned. When
\code{type = "terms"}, returns each ridge and smooth function separately.
When \code{type = "indices"}, returns predicted indices values.}

\item{select}{A numeric or character vector indicating terms to return
when \code{type = "terms"} or \code{type = "indices"}.}

\item{na.action}{A function indicating how to treat NA values when 
when \code{newdata} is not missing.}

\item{...}{For compatibility with the default \code{predict} method. Unused
at the moment.}
}
\value{
When \code{type = "response"} returns a vector of predicted response.
   When \code{type = "terms"} returns a d-column matrix where d is the 
   number of ridge and smooth terms. When \code{type = "indices"}, returns
   a p-column matrix where p is the number of created indices.
}
\description{
Uses a fitted \code{cgaim} object and computes prediction for the
   observed data or new data. Predicts the response, indices or 
   ridge functions values at the provided data.
}
\details{
When \code{newdata} is provided, it must contain all variables used in
   the fitted model.
}
