% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.cgaim}
\alias{print.cgaim}
\title{Print a cgaim object}
\usage{
\method{print}{cgaim}(x, ...)
}
\arguments{
\item{x}{A \code{cgaim} object.}

\item{...}{For compatibility with the default \code{print} method. Unused
at the moment.}
}
\description{
Default method to print the result from a call to \code{cgaim}.
}
\details{
Conveniently prints the formula, \code{beta} and \code{alpha} coefficients
   as well as the residual sum of squares.
}
